+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Release 5.5.0:
    - In `OpenId4VpHolder` remove `finalizeAuthorizationResponse()` taking in `CredentialSubmission`
    - In `OpenId4VpHolder` remove `finalizeAuthorizationResponseParameters()` taking in `CredentialSubmission`
    - In `OpenId4VpWallet` remove `finalizeAuthorizationResponseParameters()` taking in `CredentialSubmission`
+ - Update implementation of [OpenID for Verifiable Credential Issuance](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html) to draft 15:
+   - Remove functionality to request issuance of certain claims only, as this has been dropped from OpenID4VCI entirely
+   - Deprecate `WalletService.RequestOptions.requestedAttributes`
+   - Deprecate methods in `OpenId4VciClient` containing parameter for `requestedAttributes`
+   - In `OpenId4VciClient.startProvisioningWithAuthRequest()` remove parameter `requestedAttributes`
+   - In `OpenId4VciClient.loadCredentialWithOffer()` remove parameter `requestedAttributes`
 
 Release 5.4.3:
  - Fix property names for serialized RQES data classes